task 2

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -12,7 +12,7 @@
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	int mask = 1, length = 0;
+	unsigned long int mask = 1, length = 0;
 	unsigned long int m;
 
 	if (index > 64)
